# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amane API",
    "version": "0.7.2"
  },
  "paths": {
...
    "/api/libraries": {
      "get": {
        "tags": [
          "libraries"
        ],
        "summary": "List Libraries",
        "operationId": "list_libraries",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "libraries"
        ],
        "summary": "Create Library",
        "operationId": "create_library",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}": {
      "get": {
        "tags": [
          "libraries"
        ],
        "summary": "Get Library",
        "operationId": "get_library",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "libraries"
        ],
        "summary": "Update Library",
        "operationId": "update_library",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Library Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "libraries"
        ],
        "summary": "Delete Library",
        "description": "级联删除该库 MediaFile (library_id 非空 FK). 仅删除数据库索引, 不动磁盘文件.",
        "operationId": "delete_library",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Library Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadableResource": {
        "type": "string",
        "enum": [
          "thumb",
          "poster",
          "extrafanart",
          "trailer"
        ],
        "title": "DownloadableResource",
        "description": "影片附属资源类型: 刮削写入 Resource, 整理时按库配置复制到库路径."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryAutomation": {
        "type": "string",
        "enum": [
          "none",
          "watch",
          "scrape"
        ],
        "title": "LibraryAutomation",
        "description": "媒体库自动化级别. 含更低级别的行为; 自动整理尚未开放."
      },
      "LibraryCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "automation": {
            "$ref": "#/components/schemas/LibraryAutomation",
            "default": "scrape"
          },
          "recursive": {
            "type": "boolean",
            "title": "Recursive",
            "default": true
          },
          "patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Patterns",
            "default": []
          },
          "move_mode": {
            "$ref": "#/components/schemas/MoveMode",
            "default": "move"
          },
          "video_template": {
            "type": "string",
            "title": "Video Template",
            "default": "{studio}/{number}/{number}[-CD{cd?}][-{sub?}].{ext}"
          },
          "link_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Template"
          },
          "link_mode": {
            "$ref": "#/components/schemas/LinkMode",
            "default": "strm"
          },
          "strm_content_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strm Content Template"
          },
          "thumb_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumb Template"
          },
          "poster_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Template"
          },
          "fanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fanart Template"
          },
          "extrafanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extrafanart Template"
          },
          "nfo_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nfo Template"
          },
          "trailer_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailer Template"
          },
          "subtitle_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Template"
          },
          "subtitle_extensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subtitle Extensions"
          },
          "write_nfo": {
            "type": "boolean",
            "title": "Write Nfo",
            "default": true
          },
          "copy_resources": {
            "items": {
              "$ref": "#/components/schemas/DownloadableResource"
            },
            "type": "array",
            "title": "Copy Resources"
          },
          "trailer_pattern": {
            "type": "string",
            "title": "Trailer Pattern",
            "default": "(?i)trailer"
          },
          "blacklist_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blacklist Patterns",
            "default": []
          },
          "min_file_size": {
            "type": "integer",
            "title": "Min File Size",
            "default": 0
          },
          "scan": {
            "type": "boolean",
            "title": "Scan",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "LibraryCreateRequest"
      },
      "LibraryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LibraryResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LibraryListResponse"
      },
      "LibraryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "automation": {
            "$ref": "#/components/schemas/LibraryAutomation"
          },
          "recursive": {
            "type": "boolean",
            "title": "Recursive"
          },
          "patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Patterns",
            "default": []
          },
          "move_mode": {
            "$ref": "#/components/schemas/MoveMode"
          },
          "video_template": {
            "type": "string",
            "title": "Video Template"
          },
          "link_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Template"
          },
          "link_mode": {
            "$ref": "#/components/schemas/LinkMode"
          },
          "strm_content_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strm Content Template"
          },
          "thumb_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumb Template"
          },
          "poster_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Template"
          },
          "fanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fanart Template"
          },
          "extrafanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extrafanart Template"
          },
          "nfo_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nfo Template"
          },
          "trailer_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailer Template"
          },
          "subtitle_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Template"
          },
          "subtitle_extensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subtitle Extensions"
          },
          "write_nfo": {
            "type": "boolean",
            "title": "Write Nfo"
          },
          "copy_resources": {
            "items": {
              "$ref": "#/components/schemas/DownloadableResource"
            },
            "type": "array",
            "title": "Copy Resources"
          },
          "trailer_pattern": {
            "type": "string",
            "title": "Trailer Pattern"
          },
          "blacklist_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blacklist Patterns"
          },
          "min_file_size": {
            "type": "integer",
            "title": "Min File Size"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "automation",
          "recursive",
          "move_mode",
          "video_template",
          "link_mode",
          "subtitle_extensions",
          "write_nfo",
          "copy_resources",
          "trailer_pattern",
          "blacklist_patterns",
          "min_file_size"
        ],
        "title": "LibraryResponse"
      },
      "LibraryUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "automation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LibraryAutomation"
              },
              {
                "type": "null"
              }
            ]
          },
          "recursive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recursive"
          },
          "patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patterns"
          },
          "move_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoveMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "video_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Template"
          },
          "link_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Template"
          },
          "link_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "strm_content_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strm Content Template"
          },
          "thumb_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumb Template"
          },
          "poster_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Template"
          },
          "fanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fanart Template"
          },
          "extrafanart_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extrafanart Template"
          },
          "nfo_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nfo Template"
          },
          "trailer_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailer Template"
          },
          "subtitle_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Template"
          },
          "subtitle_extensions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Extensions"
          },
          "write_nfo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Write Nfo"
          },
          "copy_resources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DownloadableResource"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Copy Resources"
          },
          "trailer_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailer Pattern"
          },
          "blacklist_patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blacklist Patterns"
          },
          "min_file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min File Size"
          }
        },
        "type": "object",
        "title": "LibraryUpdateRequest"
      },
      "LinkMode": {
        "type": "string",
        "enum": [
          "strm",
          "symlink"
        ],
        "title": "LinkMode",
        "description": "整理后在 link_template 位置如何指向真实视频."
      },
...
      "MoveMode": {
        "type": "string",
        "enum": [
          "move",
          "copy",
          "hardlink",
          "symlink"
        ],
        "title": "MoveMode",
        "description": "整理时将源文件写入模板路径的方式."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}